Refactor: Extract stats table player card workflow
